Day 5 - part 2

+++ app/src/main/java/com/tilton/aoc2023/domain/model/Answer.kt
@@ -1,3 +1,3 @@
 package com.tilton.aoc2023.domain.model
 
-data class Answer(val day: Int, val part1: Int?, val part2: Int?)
+data class Answer(val day: Int, val part1: Long?, val part2: Long?)
