+++ app/src/main/java/com/tilton/aoc2023/domain/solution/Solution.kt
@@ -2,6 +2,6 @@ package com.tilton.aoc2023.domain.solution
 
 interface Solution {
     val input: List<String>
-    fun part1(): Int?
-    fun part2(): Int?
+    fun part1(): Long?
+    fun part2(): Long?
 }
